Move the boot banner out of AppRuntime.cpp
AppRuntime.cpp was 816 lines, past the ~600 the modularity rule allows, and
the serial identify command about to be added belongs with the banner it
repeats. So the banner moves first, on its own: logIdentity() and the
board/build/time-sync lines that begin() printed now live in
AppRuntimeIdentity.cpp, behind one logBootBanner() call.

No behaviour change -- the lines are moved verbatim and print in the same
order. AppRuntime.cpp is 697 lines, still over the limit; this carves off
only what the next commit touches.

Flash 2,439,461 bytes on this branch (the branch name is in the image).
