+++ src/engine/AppRuntime.cpp
@@ -1,12 +1,9 @@
 #include "AppRuntime.h"
 
-#include <esp_chip_info.h>
 #include <esp_mac.h>
 #include <esp_system.h>
 #include <math.h>
 #include "engine/NearbyPlay.h"
-#include "AppVersion.h"
-#include "BuildStamp.h"
 #include "hal/BleBeacon.h"
 #include "hal/Clock.h"
 #include "hal/Watchdog.h"
@@ -192,29 +189,7 @@ void BrainoApp::begin() {
      * engine/Entropy.h. */
     content_.begin(board_);
     Clock::begin();
-    logIdentity();
-    Serial.printf("[boot] board=%s rot=%d layout=%s (landscape=%d portrait=%d)\n",
-                  BOARD.name,
-                  (int)board_.displayRotation(),
-                  board_.layoutMode() == Board::LayoutMode::Vertical ? "Vertical" : "Horizontal",
-                  (int)BOARD.panel.landscapeRotation,
-                  (int)BOARD.panel.portraitRotation);
-    /* Beside the board line, so a serial log pasted into an issue identifies
-     * the firmware that produced it. "Which build was that?" has cost more
-     * than one round trip on a bug report. */
-    Serial.printf("[boot] build=%s built=%s version=%s\n",
-                  BuildStamp::describe(), BuildStamp::builtAt(), BRAINO_VERSION);
-    /* The saved SSID is deliberately absent from this line, and from every
-     * other Serial write in the tree. A serial log is the one artifact of
-     * this device that routinely leaves the house -- pasted into a bug
-     * report, captured by whoever plugs a cable in -- and the network's name
-     * belongs to the household, not to the firmware. `creds` answers the only
-     * question a log needs answered: whether a network is configured at all.
-     * The name itself is still on the device, in System Info and the network
-     * activity list, where the person reading it is the person holding it. */
-    Serial.printf("[boot] ntp=%d creds=%d tzmin=%d\n",
-                  (int)board_.ntpEnabled(), (int)board_.hasWifiCredentials(),
-                  (int)board_.tzOffsetMinutes());
+    logBootBanner();
     board_.beginTimeSync();
     /* After Board::begin(), which is what brings the beacon up: Nearby play
      * rides on that radio and must not try to arm itself before it exists. */
@@ -261,101 +236,6 @@ void BrainoApp::begin() {
      * generated by tickAudio() from the loop -- so the profile picker paints
      * and takes touches all the way through it. */
     board_.playSound(Sound::Boot);
-}
-
-/* WHO AM I, PRINTED BEFORE ANYTHING ELSE CAN GO WRONG.
- *
- * Four boards on one desk, COM numbers that Windows reshuffles on every
- * replug, and a boot banner that only said the board NAME -- which is compiled
- * in, so it reports which firmware is on the chip and not which panel is under
- * it. Between them those cost most of an afternoon: a 2.8-inch board was
- * flashed with the 4-inch build and reported itself as a 4-inch quite
- * happily, and a board nobody had seen before was identified from a firmware
- * image I had put there myself half an hour earlier.
- *
- * The fix is that the device says what it is, from the parts of itself that
- * cannot be wrong:
- *
- *   mac    burned into eFuse, unique per chip, and the ONLY identifier here
- *          that survives being flashed with the wrong image. This is the line
- *          that tells two identical-looking boards apart.
- *   chip   model, revision and core count, straight from esp_chip_info().
- *   flash  size as the ROM reports it, not as the profile claims.
- *   panel  the driver this binary was BUILT with, its size, rotation and
- *          backlight pin -- so a wrong-profile flash is visible in one line
- *          instead of being inferred from a photograph of odd colours.
- *   id     the display controller's own ID register, read back over MISO.
- *          This is the only fact here that comes from the PANEL rather than
- *          from the build, which makes it the one that can contradict the
- *          profile. 0x00 0x93 0x41 is an ILI9341; 0x85 0x85 0x52 is an
- *          ST7789. All-zero or all-ones means MISO is not wired back, which
- *          is common and is not itself a fault -- it just means this
- *          particular check cannot help on this board.
- *   touch  controller and its pins, since a silent panel and a dead digitiser
- *          look identical from across a room.
- *
- * The MAC is a hardware serial number, not personal data: it identifies the
- * board on the bench and never leaves the wire. No profile name, no Wi-Fi
- * credential and no player data is printed here -- see the note by the
- * Wi-Fi line below, which is the rule this has to keep. */
-void BrainoApp::logIdentity() {
-    esp_chip_info_t chip{};
-    esp_chip_info(&chip);
-
-    /* THE MAC IS DELIBERATELY NOT HERE. It was, and this line is the one that
-     * gets pasted into a public issue, so it published a permanent hardware
-     * identifier every time somebody asked for help. `device=` is generated by
-     * the firmware, means nothing off this board, and is reissued by a factory
-     * reset -- see Board::deviceId(). If you genuinely need the MAC while
-     * holding the board, esptool reads it off the chip. */
-    Serial.printf("[boot] device=%s "
-                  "chip=%s rev=%u cores=%u flash=%luKB psram=%luKB\n",
-                  board_.deviceId(),
-                  ESP.getChipModel(), (unsigned)ESP.getChipRevision(),
-                  (unsigned)chip.cores,
-                  (unsigned long)(ESP.getFlashChipSize() / 1024),
-                  (unsigned long)(ESP.getPsramSize() / 1024));
-
-    /* THERE IS NO `id=` FIELD HERE, AND THERE MUST NOT BE ONE AGAIN.
-     *
-     * 5.9.0 read the controller's ID register with three
-     * `tft.readcommand8(0x04, n)` calls and printed the answer. It shipped,
-     * and it broke the display on two of the seven boards: the panel came up
-     * with the top bar crushed into a band at the bottom edge and the rest of
-     * the screen never painted -- an address-window symptom, on a board whose
-     * loop was meanwhile running at 48fps with a 23ms worst frame, a flat heap
-     * and no stall. Every diagnostic said the firmware was healthy, because it
-     * was. The corruption was in the panel, put there by the diagnostic.
-     *
-     * TFT_eSPI's readcommand8() is not a read. It writes 0xD9 -- an
-     * undocumented index-register command -- toggles CS mid-sequence, issues
-     * the command and clocks a byte back, and it restores neither the address
-     * window nor MADCTL. On a panel that answers you get an ID; on a panel
-     * whose MISO is not wired back you get whatever the bus floats to AND a
-     * controller left mid-command, which the next write inherits.
-     *
-     * The field was worth nothing even where it appeared to work: across all
-     * seven supported boards it only ever answered 00:00:00 (MISO absent) or
-     * FF:FF:FF (floating high). Not one returned a real ID. So this was a
-     * diagnostic that could not diagnose, paid for with a corrupted panel.
-     *
-     * If a future board genuinely wires MISO back and the ID is genuinely
-     * wanted, it needs a PanelProfile field saying so and a setRotation()
-     * reissued afterwards to rebuild the window -- never an unconditional
-     * read. Until a board needs it, the correct version of this line is the
-     * one that does not touch the panel at all. */
-    Serial.printf("[boot] panel=%s%s %dx%d bl=IO%d\n",
-                  GUME_PANEL_DRIVER, GUME_PANEL_INVERTED ? "+inv" : "",
-                  (int)BOARD.panel.nativeWidth, (int)BOARD.panel.nativeHeight,
-                  (int)BOARD.panel.backlightPin);
-
-    Serial.printf("[boot] touch=%s cs=IO%d irq=IO%d sd=%s rgb=%s bat=%s\n",
-                  BOARD.touch.kind == TouchKind::CapacitiveFt6336u
-                      ? "FT6336U/i2c" : "XPT2046/bitbang",
-                  (int)BOARD.touch.cs, (int)BOARD.touch.irq,
-                  BOARD.hasSdSlot() ? "yes" : "no",
-                  BOARD.hasRgbLed() ? "yes" : "no",
-                  BOARD.hasBatterySense() ? "yes" : "no");
 }
 
 void BrainoApp::loop() {
